+++ test/e2e/operator.go
@@ -27,33 +27,6 @@ import (
 	"github.com/Azure/ARO-RP/pkg/util/ready"
 )
 
-func updatedObjects(ctx context.Context, nsfilter string) ([]string, error) {
-	pods, err := clients.Kubernetes.CoreV1().Pods("openshift-azure-operator").List(ctx, metav1.ListOptions{
-		LabelSelector: "app=aro-operator-master",
-	})
-	if err != nil {
-		return nil, err
-	}
-	if len(pods.Items) != 1 {
-		return nil, fmt.Errorf("%d aro-operator-master pods found", len(pods.Items))
-	}
-	b, err := clients.Kubernetes.CoreV1().Pods("openshift-azure-operator").GetLogs(pods.Items[0].Name, &corev1.PodLogOptions{}).DoRaw(ctx)
-	if err != nil {
-		return nil, err
-	}
-
-	rx := regexp.MustCompile(`msg="(Update|Create) ([-a-zA-Z/.]+)`)
-	changes := rx.FindAllStringSubmatch(string(b), -1)
-	result := make([]string, 0, len(changes))
-	for _, change := range changes {
-		if nsfilter == "" || strings.Contains(change[2], "/"+nsfilter+"/") {
-			result = append(result, change[1]+" "+change[2])
-		}
-	}
-
-	return result, nil
-}
-
 var _ = Describe("ARO Operator - Internet checking", func() {
 	var originalURLs []string
 	BeforeEach(func() {
@@ -131,7 +104,9 @@ var _ = Describe("ARO Operator - Geneva Logging", func() {
 		}
 
 		err := wait.PollImmediate(30*time.Second, 15*time.Minute, mdsdReady)
-		Expect(err).NotTo(HaveOccurred())
+		if err != nil {
+			clients.Kubernetes.EventsV1().
+		}
 		initial, err := updatedObjects(context.Background(), "openshift-azure-logging")
 		Expect(err).NotTo(HaveOccurred())
